requirements.txt
streamlit==1.18.0
requests==2.28.0
llama-index==0.5.2
lizard==1.3.1
radon==5.1.0
gitpython==3.1.32
chardet==5.1.0
pandas==2.1.0
plotly==5.11.0